cn10k: mmc: Fix ADMA transfer limits
Change default block count and block max size.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/u-boot/+/97293
Tested-by: Chandrakala Chavva <cchavva@marvell.com>

+++ drivers/mmc/sdhci-adma.c
@@ -48,7 +48,7 @@ void sdhci_prepare_adma_table(struct sdhci_adma_desc *table,
 	int i = desc_count;
 
 	while (--i) {
-		sdhci_adma_desc(desc, addr, ADMA_MAX_LEN, false);
+		sdhci_adma_desc(desc, addr, (u16)ADMA_MAX_LEN, false);
 		addr += ADMA_MAX_LEN;
 		trans_bytes -= ADMA_MAX_LEN;
 		desc++;
